+++ app/models/document_models.py
@@ -28,13 +28,17 @@ class Document(Base):
     mimetype = Column(String(256), index=True, nullable=False)
 
     thumbnail_filename = Column(String(128), nullable=True, unique=True)
+
     comments_count = Column(Integer, index=True, default=0)
     downloads_count = Column(Integer, index=True, default=0)
+    favorites_count = Column(Integer, index=True, default=0)
 
     document_user = relationship(
         "User", back_populates="user_documents", lazy="joined")
+
     document_collection = relationship(
         "Collection", back_populates="collection_documents", lazy="joined")
+
     document_tags = relationship(
         "Tag", back_populates="tag_document", lazy="joined",
         cascade="all, delete-orphan")
@@ -45,6 +49,10 @@ class Document(Base):
 
     document_downloads = relationship(
         "Download", back_populates="download_document", lazy="noload",
+        cascade="all, delete-orphan")
+
+    document_favorites = relationship(
+        "Favorite", back_populates="favorite_document", lazy="noload",
         cascade="all, delete-orphan")
 
     def __init__(self, user_id: int, collection_id: int, document_name: str,
@@ -63,6 +71,7 @@ class Document(Base):
         self.thumbnail_filename = thumbnail_filename
         self.comments_count = 0
         self.downloads_count = 0
+        self.favorites_count = 0
 
     @property
     def file_path(self):
@@ -96,5 +105,6 @@ class Document(Base):
             "thumbnail_url": self.thumbnail_url,
             "comments_count": self.comments_count,
             "downloads_count": self.downloads_count,
+            "favorites_count": self.favorites_count,
             "document_tags": self.tag_values,
         }